# ddl: speed up the operation of "drop database"