<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #16983 <!-- REMOVE this line if no issue to close -->

Problem Summary:
<img width="1915" alt="Screen Shot 2020-05-07 at 12 42 14" src="https://user-images.githubusercontent.com/4242506/81255646-c5eb1700-9060-11ea-9ecc-0a34278bcdc4.png">
Looking at the logs, this drop schema operation really took almost 1 hour. 
![WeChatWorkScreenshot_b015a9fc-b256-47d1-82e6-8493c00d7890](https://user-images.githubusercontent.com/4242506/81256736-f54f5300-9063-11ea-8a7d-b338bb270d20.png)
The insert statements take more than 50 minutes. The other time consuming is due to the synchronization method of waiting for 2 * lease.
This PR we speed up the insert statements in "drop schema" operation.

### What is changed and how it works?

What's Changed:
In the above case, the database has 1587204 tables, so it needs to execute 1587204 insert statements. In this PR we do batch insertion.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test (original tests, and set `BatchInsertDeleteRangeSize` to 2)
- Manual test (add detailed scripts or steps below)

Using TiUP to start a cluster for 3 tikv，1 tidb and 1 pd.

dbx* has 1 table, dby* has 100 tables, dbz* has 10000 tables.

Before this PR:
```
mysql> drop database dbx1;
Query OK, 0 rows affected (0.33 sec)
mysql> drop database dby1;
Query OK, 0 rows affected (0.69 sec)
mysql> drop database dbz1;
Query OK, 0 rows affected (55.89 sec)
```
After this PR:
```
mysql> drop database dbx0;
Query OK, 0 rows affected (0.27 sec)
mysql> drop database dby0;
Query OK, 0 rows affected (0.38 sec)
mysql> drop database dbz0;
Query OK, 0 rows affected (3.77 sec)
```

### Release note <!-- bugfixes or new feature need a release note -->

- Speed up the deletion of databases with a large number of tables