@bb7133 2020-05-09
Failed cases:

```
FAIL: statistics_handler_test.go:86: testDumpStatsSuite.TestDumpStatsAPI

[2020-05-09T06:09:38.384Z] 

[2020-05-09T06:09:38.384Z] statistics_handler_test.go:87:

[2020-05-09T06:09:38.384Z]     ds.startServer(c)

[2020-05-09T06:09:38.384Z] statistics_handler_test.go:64:

[2020-05-09T06:09:38.384Z]     c.Assert(err, IsNil)

[2020-05-09T06:09:38.384Z] ... value *errors.withStack = listen tcp 0.0.0.0:10092: bind: address already in use ("listen tcp 0.0.0.0:10092: bind: address already in use")
```
https://internal.pingcap.net/idc-jenkins/blue/organizations/jenkins/tidb_ghpr_unit_test/detail/tidb_ghpr_unit_test/35674/